requirement.txt
anyio==3.3.4
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
distlib==0.3.3
Django==3.2.9
djangorestframework==3.12.4
fastapi==0.70.0
filelock==3.3.2
h11==0.12.0
idna==3.3
lxml==4.6.4
nsepy==0.8
numpy==1.21.4
pandas==1.3.4
platformdirs==2.4.0
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
sqlparse==0.4.2
starlette==0.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
virtualenv==20.10.0
virtualenv-clone==0.5.7
watchgod==0.7
websockets==10.0
